Fix upgrade-database for configfiles
We are not storing schema version information for configfiles, since we won't
be able to use them and the user is responsible for updating them.
With this patch, we tell upgrade-db to continue with the rest of the databases
if any of them is a configfile, rather than abandoning the entire process.

Signed-off-by: Patrick Uiterwijk <puiterwijk@redhat.com>
Reviewed-by: Pierre-Yves Chibon <pingou@pingoured.fr>