Add compatibility with jwcrypto 0.3.0 and higher
Jwcrypto 0.3.0 has moved the actual set with keys into a dict field 'keys' rather than
being the set itself.
We can't use the export function because that exports the private keys by default, and
there's no way in 0.2.0 to indicate we don't want private keys, and we can't determine
whether the function will accept this argument.

Signed-off-by: Patrick Uiterwijk <puiterwijk@redhat.com>
Reviewed-by: Howard Johnson <merlin@merlinthp.org>