Styled the user consent page for openID Connect
This styles up the user consent page for openID Connect. It presents the
list of items to consent centered in the middle of a fluid full-page
layout, with the user action to accept or reject at the bottom of the
list

There is a new master-user template that this page now uses to display
the content, as the the master template uses the login-pf layout that
does not work well with long lists of content.

This is mostly Ryan Lerch's work from PR#96, with some minor CSS and
HTML changes to make it consistent with the login page.

Merges: #96
Signed-off-by: Ryan Lerch <rlerch@redhat.com>
Signed-off-by: Howard Johnson <merlin@merlinthp.org>
Reviewed-by: Patrick Uiterwijk <puiterwijk@redhat.com>