Retain transaction ID through 401 pages
If we get a 401 during a transaction (i.e. failing authentication),
ipsilon shows the 401 error page, which has a link back to the login
page.  Clicking this link causes a new transaction to be created,
losing any in-progress service provider authentication, and not
redirecting back to the SP after a successful transaction.

If we hit the 401 page and there's a current transaction, add the
transaction ID to the link back to the login page, so the current
transaction is preserved.

Merges: #121
Signed-off-by: Howard Johnson <merlin@merlinthp.org>
Reviewed-by: Patrick Uiterwijk <puiterwijk@redhat.com>