Mark lines as noqa to avoid pep8 E126 on EL7
These PEP-8 errors are caused by an extra indent for dict values.
We would like to do indent these extra to indicate they are not just another
key in the dict, but pep8 on EL7 does not accept this style, because it thinks
that we are in a function definition.

Let's just ignore these couple of lines for now.

Signed-off-by: Patrick Uiterwijk <puiterwijk@redhat.com>
Reviewed-by: Howard Johnson <merlin@merlinthp.org>