Abort OpenId responses with headers over 8kb
They will not be able to be returned through Apache, and if we don't
error this way, Apache will log obscure things.

This will happen if a user for example has 3 16KB SSH keys (live
example seen), and they are all requested.
OpenID 2.0 wants to encode the entire response in an HTTP Location
header, but when Apache sees this in the headers, it thinks we
made a mistake, and aborts the response.

Ticket: #238
Merges: #168
Signed-off-by: Patrick Uiterwijk <puiterwijk@redhat.com>
Reviewed-by: Howard Johnson <merlin@merlinthp.org>